Conditional Functions
This Python Script contains three simple functions, each with specific logic. Function 1 adds two numbers together, and if they are the same, it multiplies them. Function 2 returns either the value of A if B is true or double the value of A if B is false. Function 3 returns both if A and B are true, returns true if A is true and B is false, and false in any other cases.